# Conflicts:
#	Jenkinsfile
#	parent-poms/parent/pom.xml

+++ dataflow-service/src/main/java/org/eea/dataflow/controller/ContributorControllerImpl.java
@@ -2,11 +2,16 @@ package org.eea.dataflow.controller;
 
 import java.util.List;
 import org.eea.dataflow.service.ContributorService;
+import org.eea.dataflow.service.impl.DataflowServiceImpl;
+import org.eea.exception.EEAErrorMessage;
 import org.eea.exception.EEAException;
 import org.eea.interfaces.controller.dataflow.ContributorController;
 import org.eea.interfaces.controller.ums.UserManagementController.UserManagementControllerZull;
 import org.eea.interfaces.vo.contributor.ContributorVO;
-import org.eea.interfaces.vo.ums.UserRepresentationVO;
+import org.eea.interfaces.vo.dataflow.enums.TypeDataflowEnum;
+import org.eea.interfaces.vo.enums.EntityClassEnum;
+import org.eea.interfaces.vo.ums.enums.SecurityRoleEnum;
+import org.eea.security.authorization.ObjectAccessRoleEnum;
 import org.eea.utils.LiteralConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,6 +20,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -30,11 +36,13 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import springfox.documentation.annotations.ApiIgnore;
 
 /**
  * The Class ContributorControllerImpl.
  */
 @RestController
+@ApiIgnore
 @Api(tags = "Contributors : Contributors Manager")
 @RequestMapping("/contributor")
 public class ContributorControllerImpl implements ContributorController {
@@ -67,6 +75,10 @@ public class ContributorControllerImpl implements ContributorController {
   @Autowired
   private UserManagementControllerZull userManagementControllerZull;
 
+  /** The dataflow service. */
+  @Autowired
+  private DataflowServiceImpl dataflowService;
+
   /**
    * Delete requester.
    *
@@ -76,23 +88,27 @@ public class ContributorControllerImpl implements ContributorController {
   @Override
   @PreAuthorize("secondLevelAuthorize(#dataflowId,'DATAFLOW_STEWARD','DATAFLOW_CUSTODIAN') || hasRole('ADMIN')")
   @DeleteMapping(value = "/requester/dataflow/{dataflowId}")
-  @ApiOperation(value = "Delete one requester in a Dataflow")
+  @ApiOperation(value = "Delete one requester in a Dataflow", hidden = true)
   @ApiResponses(value = {@ApiResponse(code = 200, message = "Successfully deleted requester"),
       @ApiResponse(code = 204, message = "Successfully deleted requester"),
       @ApiResponse(code = 401, message = "You are not authorized to view the resource"),
       @ApiResponse(code = 404, message = "The email doesn't exist in Repornet"),
       @ApiResponse(code = 500, message = "Internal Server Error")})
-  @ApiParam()
   public void deleteRequester(
       @ApiParam(value = "Dataflow Id", example = "0") @PathVariable("dataflowId") Long dataflowId,
       @ApiParam(type = "Object",
           value = "Contributors Properties") @RequestBody ContributorVO contributorVO) {
+    // check permission not allowed for custodian in bdr dataflow
+    checkIsBussinesCustodian(dataflowId);
     // we can only remove role of editor, reporter or reporter partition type
     try {
-      checkAccount(dataflowId, contributorVO.getAccount());
-      contributorService.deleteContributor(dataflowId, contributorVO.getAccount(),
-          contributorVO.getRole(), null);
-      LOG.info("requester {} Deleted", contributorVO.getAccount());
+      if (checkAccount(dataflowId, contributorVO.getAccount())) {
+        contributorService.deleteContributor(dataflowId, contributorVO.getAccount(),
+            contributorVO.getRole(), null);
+        LOG.info("requester {} Deleted", contributorVO.getAccount());
+      } else
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+            String.format(THE_EMAIL_NOT_EXISTS, contributorVO.getAccount()));
     } catch (EEAException e) {
       if (HttpStatus.NOT_FOUND.toString().equals(e.getMessage())) {
         throw new ResponseStatusException(HttpStatus.NOT_FOUND,
@@ -114,7 +130,7 @@ public class ContributorControllerImpl implements ContributorController {
   @Override
   @PreAuthorize("secondLevelAuthorize(#dataflowId,'DATAFLOW_STEWARD','DATAFLOW_CUSTODIAN', 'DATAFLOW_LEAD_REPORTER')")
   @DeleteMapping(value = "/reporter/dataflow/{dataflowId}/provider/{dataProviderId}")
-  @ApiOperation(value = "Delete one Reporter in a Dataflow")
+  @ApiOperation(value = "Delete one Reporter in a Dataflow", hidden = true)
   @ApiResponses(value = {@ApiResponse(code = 200, message = "Successfully deleted reporter"),
       @ApiResponse(code = 204, message = "Successfully deleted reporter"),
       @ApiResponse(code = 401, message = "You are not authorized to view the resource"),
@@ -129,10 +145,18 @@ public class ContributorControllerImpl implements ContributorController {
           value = "Contributors Properties") @RequestBody ContributorVO contributorVO) {
     // we can only remove role of editor, reporter or reporter partition type
     try {
-      checkAccount(dataflowId, contributorVO.getAccount());
-      contributorService.deleteContributor(dataflowId, contributorVO.getAccount(),
-          contributorVO.getRole(), dataProviderId);
-      LOG.info("Reporter {} Deleted", contributorVO.getAccount());
+      if (checkAccount(dataflowId, contributorVO.getAccount())) {
+        contributorService.deleteContributor(dataflowId, contributorVO.getAccount(),
+            contributorVO.getRole(), dataProviderId);
+        LOG.info("Reporter {} Deleted", contributorVO.getAccount());
+      } else if (contributorService.findTempUserByAccountAndDataflow(contributorVO.getAccount(),
+          dataflowId, dataProviderId) != null) {
+        contributorService.deleteTemporaryUser(dataflowId, contributorVO.getAccount(),
+            contributorVO.getRole(), dataProviderId);
+        LOG.info("Temporary Reporter {} Deleted", contributorVO.getAccount());
+      } else
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+            String.format(THE_EMAIL_NOT_EXISTS, contributorVO.getAccount()));
     } catch (EEAException e) {
       if (HttpStatus.NOT_FOUND.toString().equals(e.getMessage())) {
         throw new ResponseStatusException(HttpStatus.NOT_FOUND,
@@ -157,7 +181,7 @@ public class ContributorControllerImpl implements ContributorController {
       produces = MediaType.APPLICATION_JSON_VALUE)
   @ApiOperation(value = "Find all Requesters in a Dataflow",
       produces = MediaType.APPLICATION_JSON_VALUE, response = ContributorVO.class,
-      responseContainer = "List")
+      responseContainer = "List", hidden = true)
   public List<ContributorVO> findRequestersByGroup(@ApiParam(type = "Long", value = "Dataflow Id",
       example = "0") @PathVariable("dataflowId") Long dataflowId) {
     // we can find requesters,
@@ -179,15 +203,23 @@ public class ContributorControllerImpl implements ContributorController {
   @PreAuthorize("secondLevelAuthorize(#dataflowId,'DATAFLOW_STEWARD','DATAFLOW_CUSTODIAN', 'DATAFLOW_LEAD_REPORTER')")
   @ApiOperation(value = "Find all Reporters in a Dataflow",
       produces = MediaType.APPLICATION_JSON_VALUE, response = ContributorVO.class,
-      responseContainer = "List")
+      responseContainer = "List", hidden = true)
   public List<ContributorVO> findReportersByGroup(
       @ApiParam(type = "Long", value = "Dataflow Id",
           example = "0") @PathVariable("dataflowId") Long dataflowId,
       @ApiParam(type = "Long", value = "Dataprovider Id",
-          example = "0") @PathVariable("dataproviderId") Long dataproviderId) {
+          example = "0") @PathVariable("dataproviderId") Long dataProviderId) {
     // find reporters or reporter partition roles based on the dataflow state
-    return contributorService.findContributorsByResourceId(dataflowId, dataproviderId,
-        LiteralConstants.REPORTER);
+    List<ContributorVO> tempReporterWrite = contributorService.findTempUserByRoleAndDataflow(
+        SecurityRoleEnum.REPORTER_WRITE.toString(), dataflowId, dataProviderId);
+    List<ContributorVO> tempReporterRead = contributorService.findTempUserByRoleAndDataflow(
+        SecurityRoleEnum.REPORTER_READ.toString(), dataflowId, dataProviderId);
+    List<ContributorVO> reportersList = contributorService.findContributorsByResourceId(dataflowId,
+        dataProviderId, LiteralConstants.REPORTER);
+    reportersList.addAll(tempReporterWrite);
+    reportersList.addAll(tempReporterRead);
+
+    return reportersList;
   }
 
   /**
@@ -204,7 +236,7 @@ public class ContributorControllerImpl implements ContributorController {
   @PutMapping(value = "/requester/dataflow/{dataflowId}",
       produces = MediaType.APPLICATION_JSON_VALUE)
   @ApiOperation(value = "Update one Requester in a Dataflow",
-      produces = MediaType.APPLICATION_JSON_VALUE, response = ResponseEntity.class)
+      produces = MediaType.APPLICATION_JSON_VALUE, response = ResponseEntity.class, hidden = true)
   @ApiResponses(value = {@ApiResponse(code = 200, message = "Successfully update requester"),
       @ApiResponse(code = 204, message = "Successfully updated requester"),
       @ApiResponse(code = 401, message = "You are not authorized to view the resource"),
@@ -215,14 +247,19 @@ public class ContributorControllerImpl implements ContributorController {
           example = "0") @PathVariable("dataflowId") Long dataflowId,
       @ApiParam(type = "Object",
           value = "Contributors Properties") @RequestBody ContributorVO contributorVO) {
+    // check permission not allowed for custodian in bdr dataflow
+    checkIsBussinesCustodian(dataflowId);
     // we can only update an editor, reporter or reporter partition role
     // mock
     String message = "";
     HttpStatus status = HttpStatus.OK;
     try {
-      checkAccount(dataflowId, contributorVO.getAccount());
-      contributorService.updateContributor(dataflowId, contributorVO, null);
-      LOG.info("requester {} Updated", contributorVO.getAccount());
+      if (checkAccount(dataflowId, contributorVO.getAccount())) {
+        contributorService.updateContributor(dataflowId, contributorVO, null);
+        LOG.info("requester {} Updated", contributorVO.getAccount());
+      } else
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+            String.format(THE_EMAIL_NOT_EXISTS, contributorVO.getAccount()));
     } catch (EEAException e) {
       if (HttpStatus.NOT_FOUND.toString().equals(e.getMessage())) {
         message = String.format(THE_EMAIL_NOT_EXISTS, contributorVO.getAccount());
@@ -237,6 +274,7 @@ public class ContributorControllerImpl implements ContributorController {
     }
     return new ResponseEntity<>(message, status);
   }
+
 
   /**
    * Update reporter.
@@ -253,7 +291,7 @@ public class ContributorControllerImpl implements ContributorController {
   @PutMapping(value = "/reporter/dataflow/{dataflowId}/provider/{dataProviderId}",
       produces = MediaType.APPLICATION_JSON_VALUE)
   @ApiOperation(value = "Update one Reporter in a Dataflow",
-      produces = MediaType.APPLICATION_JSON_VALUE, response = ResponseEntity.class)
+      produces = MediaType.APPLICATION_JSON_VALUE, response = ResponseEntity.class, hidden = true)
   @ApiResponses(value = {@ApiResponse(code = 200, message = "Successfully update reporter"),
       @ApiResponse(code = 204, message = "Successfully updated reporter"),
       @ApiResponse(code = 401, message = "You are not authorized to view the resource"),
@@ -270,9 +308,20 @@ public class ContributorControllerImpl implements ContributorController {
     String message = "";
     HttpStatus status = HttpStatus.OK;
     try {
-      checkAccount(dataflowId, contributorVO.getAccount());
-      contributorService.updateContributor(dataflowId, contributorVO, dataProviderId);
-      LOG.info("Reporter {} Updated", contributorVO.getAccount());
+      if (checkAccount(dataflowId, contributorVO.getAccount())) {
+        contributorService.updateContributor(dataflowId, contributorVO, dataProviderId);
+        LOG.info("Reporter {} Updated", contributorVO.getAccount());
+      } else if (contributorService.findTempUserByAccountAndDataflow(contributorVO.getAccount(),
+          dataflowId, dataProviderId) != null) {
+        contributorService.updateTemporaryUser(dataflowId, contributorVO, dataProviderId);
+      } else if (contributorService.findTempUserByAccountAndDataflow(contributorVO.getAccount(),
+          dataflowId, dataProviderId) == null) {
+        contributorService.createTempUser(dataflowId, contributorVO, dataProviderId);
+        LOG.info("Inserting user {} into temp user table", contributorVO.getAccount());
+      } else {
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+            "There's already a user assigned with the same e-mail to this dataflow");
+      }
     } catch (EEAException e) {
       if (HttpStatus.NOT_FOUND.toString().equals(e.getMessage())) {
         message = String.format(THE_EMAIL_NOT_EXISTS, contributorVO.getAccount());
@@ -284,6 +333,43 @@ public class ContributorControllerImpl implements ContributorController {
       LOG_ERROR.error("Error update the reporter {}.in the dataflow: {}",
           contributorVO.getAccount(), dataflowId);
     }
+    return new ResponseEntity<>(message, status);
+  }
+
+  /**
+   * Updates the reporters permissions checking if they are now registered in the system.
+   *
+   * @param dataflowId the dataflow id
+   * @param dataProviderId the data provider id
+   * @param contributorVO the contributor VO
+   * @return the response entity
+   */
+  @Override
+  @PreAuthorize("secondLevelAuthorize(#dataflowId,'DATAFLOW_LEAD_REPORTER') OR hasAnyRole('ADMIN')")
+  @PutMapping(value = "/validateReporters/dataflow/{dataflowId}/provider/{dataProviderId}",
+      produces = MediaType.APPLICATION_JSON_VALUE)
+  @ApiOperation(
+      value = "Validates reporters, checking if they are already registered in the system and gives them permissions",
+      hidden = true)
+  @ApiResponse(code = 500,
+      message = "Error creating the associated permissions for requested reporter in the dataflow")
+  public ResponseEntity validateReporters(
+      @ApiParam(value = "Dataflow ID the reporter is assigned to", required = true,
+          example = "1") @PathVariable("dataflowId") Long dataflowId,
+      @ApiParam(value = "Data Provider ID the reporter is assigned to", required = true,
+          example = "1") @PathVariable("dataProviderId") Long dataProviderId) {
+    String message = "";
+    HttpStatus status = HttpStatus.OK;
+    try {
+      contributorService.validateReporters(dataflowId, dataProviderId, true);
+    } catch (EEAException e) {
+      LOG_ERROR.error(
+          "Error creating the associated permissions for requested reporter in the dataflow: {} with data provider ID {}",
+          dataflowId, dataProviderId);
+      message = e.getMessage();
+      status = HttpStatus.BAD_REQUEST;
+    }
+
     return new ResponseEntity<>(message, status);
   }
 
@@ -323,16 +409,36 @@ public class ContributorControllerImpl implements ContributorController {
    *
    * @throws EEAException
    */
-  private void checkAccount(Long dataflowId, String account) throws EEAException {
-    // check if the email is correct
-    UserRepresentationVO emailUser = userManagementControllerZull.getUserByEmail(account);
-    if (null == emailUser || null == emailUser.getEmail()
-        || !emailUser.getEmail().equalsIgnoreCase(account)) {
-      LOG_ERROR.error(
-          "Error creating contributor with the account: {} in the dataflow {} because the email doesn't exist in the system",
-          account, dataflowId);
-      throw new EEAException(HttpStatus.NOT_FOUND.toString());
+  private boolean checkAccount(Long dataflowId, String account) {
+
+    return userManagementControllerZull.getUserByEmail(account) != null;
+  }
+
+
+  /**
+   * Check is bussines custodian.
+   *
+   * @param dataflowId the dataflow id
+   */
+  private void checkIsBussinesCustodian(Long dataflowId) {
+    if (dataflowService.isDataflowType(TypeDataflowEnum.BUSINESS, EntityClassEnum.DATAFLOW,
+        dataflowId) && !isAdminOrSteward(dataflowId)) {
+      throw new ResponseStatusException(HttpStatus.FORBIDDEN, EEAErrorMessage.FORBIDDEN);
     }
+  }
+
+  /**
+   * Checks if is admin or steward.
+   *
+   * @param dataflowId the dataflow id
+   * @return true, if is admin or steward
+   */
+  private boolean isAdminOrSteward(Long dataflowId) {
+    String roleSteward = ObjectAccessRoleEnum.DATAFLOW_STEWARD.getAccessRole(dataflowId);
+    String roleAdmin = "ROLE_" + SecurityRoleEnum.ADMIN;
+    return SecurityContextHolder.getContext().getAuthentication().getAuthorities().stream()
+        .anyMatch(role -> roleSteward.equals(role.getAuthority())
+            || roleAdmin.equals(role.getAuthority()));
   }
 
 }
